requirements.txt
pymongo==3.11.4
python-telegram-bot==13.5
dnspython==2.1.0